+++ src/main/module1/week2/task3/b/MergeSort.java
@@ -1,0 +1,61 @@
+package main.module1.week2.task3.b;
+
+import java.util.Scanner;
+
+public class MergeSort {
+    public static void mergeSort(int[] a, int n) {
+        if (n < 2) {
+            return;
+        }
+        int mid = n / 2;
+        int[] l = new int[mid];
+        int[] r = new int[n - mid];
+
+        System.arraycopy(a, 0, l, 0, mid);
+        if (n - mid >= 0) System.arraycopy(a, mid, r, 0, n - mid);
+        mergeSort(l, mid);
+        mergeSort(r, n - mid);
+
+        merge(a, l, r, mid, n - mid);
+    }
+
+    public static void merge(int[] a, int[] l, int[] r, int left, int right) {
+
+        int i = 0, j = 0, k = 0;
+        while (i < left && j < right) {
+            if (l[i] <= r[j]) {
+                a[k++] = l[i++];
+            }
+            else {
+                a[k++] = r[j++];
+            }
+        }
+        while (i < left) {
+            a[k++] = l[i++];
+        }
+        while (j < right) {
+            a[k++] = r[j++];
+        }
+    }
+
+    // A utility function to print array of size n
+    static void printArray(int[] arr) {
+        for (int j : arr) System.out.print(j + " ");
+        System.out.println();
+    }
+
+    public static void main(String[] args) {
+        Scanner s = new Scanner(System.in);
+        System.out.println("Enter the length of the array:");
+        int length = s.nextInt();
+        int[] arrayNumber = new int[length];
+
+        System.out.println("Enter the elements of the array:");
+        for (int i = 0; i < length; i++) {
+            arrayNumber[i] = s.nextInt();
+        }
+
+        MergeSort.mergeSort(arrayNumber, arrayNumber.length);
+        printArray(arrayNumber);
+    }
+}
