+++ src/main/module1/week4/task3/WrongSensorPackageException.java
@@ -1,0 +1,7 @@
+package main.module1.week4.task3;
+
+public class WrongSensorPackageException extends RuntimeException {
+    public WrongSensorPackageException(String message) {
+        super(message);
+    }
+}
