--- src/week4/task3/WrongSensorPackageException.java
@@ -1,7 +1,0 @@
-package week4.task3;
-
-public class WrongSensorPackageException extends RuntimeException {
-    public WrongSensorPackageException(String message) {
-        super(message);
-    }
-}
